Tests/LibWeb/Text/expected/UIEvents/click-event-on-checkbox-inside-label-element-not-duplicated.txt

click event on input#radio1
click event on input#radio2
